Employee and translation: fix error on delete item with Learning support team related and fix translation on delete

+++ app/src/Backoffice/Employee/Application/Delete/EmployeeDeleter.php
@@ -6,6 +6,7 @@ namespace App\Backoffice\Employee\Application\Delete;
 
 use App\Backoffice\Employee\Application\Get\Single\EmployeeFinder;
 use App\Backoffice\Employee\Domain\EmployeeRepository;
+use App\Backoffice\Employee\Domain\Exception\UnableDeleteManagerWithAssociatedLearningSupportTeam;
 
 final class EmployeeDeleter
 {
@@ -21,6 +22,10 @@ final class EmployeeDeleter
     public function __invoke(string $id)
     {
         $employee = $this->finder->__invoke($id);
+
+        if ($employee->servesAsManager()) {
+            throw new UnableDeleteManagerWithAssociatedLearningSupportTeam();
+        }
 
         $this->repository->delete($employee);
     }
